pyproject.toml
[tool.poetry]
name = "dashboard-python"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.27.0"
python-dateutil = "^2.9.0.post0"
icalendar = "^5.0.13"
soco = "^0.30.4"
fastapi = {extras = ["standard"], version = "^0.114.2"}
pyatmo = "^8.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
